feat: routings to pages using react-router-dom
We install the package of react-router-dom for using Switch and Route components to the url address routing. Because the need of access all the routings of the application, we use the  import {BrowserRouter as Router} from 'react-router-dom'  to the index.js  Then we wrap the App component with Router component at the rendering. So that, application have access to all routes and we gona use the Switch and Route components at the  App.js

With this at the rendering of the App.js inside the Switch component
  <Route path="/" component={ProductList}></Route>
we have the ProductList home page. Also we add inside all the routings of the componetnts.

For only display the home page on the "/" route we use:
  <Route exact path="/" component={ProductList} />